Fix images
The images are currently just little, unimpressive images. I adjusted the code to display the medium sized image and upon click it loads the larger version for those who want to get a better view. I just added the lightbox + colorbox plugin to my site so when you click on the beer it zooms. Here is a sample: http://hamricks.wpengine.com/test/